requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
sqlparse==0.5.3
tzdata==2025.2




admin : admin-oc
password : password-oc